Abstract out parser into its own file. Create wrapper classes around parser and group
